performance fix, dromajo SHA/version in STF output
This is version 2.0.0, there are changes that are visible to the command
line and changes to behavior.

For previous behavior this SHA is recommended:
    ff13255a50bd1b5e7597f3bf2ceaf24b782e8b72

this commit combines previous commit db4c2067982 into head.

Improve performance under linux when not tracing.
    This restores previous performance but also reduces
    some feature support.

add SHA and dromajo version to STF output for tracking

Switches have changed, for Condor use, changes are transparent

New switch:

stf_force_zero_sha : clear the SHA and version fields in STF
                     simplifies diff operations in the tests

Other switch changes:

    --stf_tracepoint ALWAYS ENABLED IN THIS VERSION
        trace macros trigger STF record generation, full file
        tracing is DISABLE IN THIS VERSION
    --stf_include_tracepoints DISABLED IN THIS VERSION
        future feature, will replace the previous --stf_include_stop_tracepoint
    --stf_trace_register_state DISABLED IN THIS VERSION
        future feature, will replace previous --stf_essential_mode
    --stf_exit_on_stop_opc DISABLED IN THIS VERSION
        this feature got eliminated in a previous commit, it has
        not been restored yet.