fix memory records bug in dromajo, stan'd regress
CN-45 bug fix candidate - more testing to be done.

this changes trace_(dread/write) to flag virtual accesses
so they are traced, and not the physicall accesses.

integrate stan'd regress ctest additions

modified some of the source in ctest to document
the load/store types that should be added.

updated the golden stf for zero SHA