implemented A* for coremark and dhrystone
See ASTAR_STATUS for details. I believe this is all
the instructions referenced by coremark and dhrystone
with a few extras in the gp-implied load/store sizes

BEQC/BNEC have a bug, the PC is modified unexpectedly
BFOS/BFOZ have an implementation but likely have bugs
Some level of interpretation required for the Andes
pseudo code on edge conditions, msb or lsb== 0, etc.